requirements.txt
fastapi
asyncio
httpx
python-dotenv
uvicorn
gunicorn